move main logic into own dir
