+++ src/helpers/index.types.ts
@@ -1,3 +1,2 @@
 export type * from './Http';
-export type * from './spotifyAxios';
 export type * from './getAuthorizationUrl';
